refactor: move isEdrSupportedChainType helper to internal/edr/chain-type
