+++ v-next/hardhat/src/internal/edr/chain-type.ts
@@ -1,10 +1,10 @@
-import type { ChainType } from "../../../../../types/network.js";
+import type { ChainType } from "../../types/network.js";
 
 import {
   GENERIC_CHAIN_TYPE,
   L1_CHAIN_TYPE,
   OPTIMISM_CHAIN_TYPE,
-} from "../../../../constants.js";
+} from "../constants.js";
 
 export function isEdrSupportedChainType(
   chainType: unknown,
